requirements.txt
requests>=2.28.0
urllib3>=1.26.0
python-dotenv>=0.19.0
